[DAPHNE-#567,#647] Adding AUTO option to calculate a grain size for SS
This pull request includes the implementation for the automated task grain size selection in DAPHNE. Here we add a new option for partitioning strategy in DAPHNE called "AUTO". When a user selects AUTO, DAPHNE will automatically calculate a fixed task grain size which frequently achieves high performance.

This is a part of #567 submitted in PR #647

Closes #647